+++ common/src/main/java/com/gitlab/controllers/api/rest/ProductSearchRestApi.java
@@ -1,10 +1,7 @@
 package com.gitlab.controllers.api.rest;
 
 import com.gitlab.dto.ProductDto;
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
+import io.swagger.annotations.*;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,5 +19,18 @@ public interface ProductSearchRestApi {
             @ApiResponse(code = 200, message = "Products found"),
             @ApiResponse(code = 204, message = "Products not present")}
     )
-    ResponseEntity<List<ProductDto>> search(@RequestParam("name") String name) throws InterruptedException;
+    ResponseEntity<List<ProductDto>> search(
+            @RequestParam("name") String name) throws InterruptedException;
+
+    @GetMapping("/api/search/paginate")
+    @ApiOperation(value = "Search products by Product.name with pagination")
+    @ApiResponses(value = {
+            @ApiResponse(code = 200, message = "Products found"),
+            @ApiResponse(code = 204, message = "Products not present")}
+    )
+    ResponseEntity<List<ProductDto>> searchPaginate(
+            @ApiParam(name = "name") @RequestParam(required = false, defaultValue = "", value = "name") String name,
+            @ApiParam(name = "page") @RequestParam(required = false, defaultValue = "0", value = "page") Integer page,
+            @ApiParam(name = "size") @RequestParam(required = false, defaultValue = "10", value = "size") Integer size)
+            throws InterruptedException;
 }