+++ gazon-service/src/main/java/com/gitlab/controller/ProductSearchController.java
@@ -4,6 +4,7 @@ import com.gitlab.controllers.api.rest.ProductSearchRestApi;
 import com.gitlab.dto.ProductDto;
 import com.gitlab.service.ProductService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.*;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -19,10 +20,13 @@ public class ProductSearchController implements ProductSearchRestApi {
     @Override
     public ResponseEntity<List<ProductDto>> search(String name) throws InterruptedException {
         List<ProductDto> products = productService.findByNameIgnoreCaseContaining(name);
+        return products.isEmpty() ? ResponseEntity.noContent().build() : ResponseEntity.ok(products);
+    }
 
-
-        return !products.isEmpty() ?
-                ResponseEntity.ok(products) :
-        ResponseEntity.noContent().build();
+    @Override
+    public ResponseEntity<List<ProductDto>> searchPaginate(String name, Integer page, Integer size) throws InterruptedException {
+        Pageable pageable = PageRequest.of(page, size);
+        List<ProductDto> products = productService.findByNameIgnoreCaseContaining(name, pageable).getContent();
+        return products.isEmpty() ? ResponseEntity.noContent().build() : ResponseEntity.ok(products);
     }
 }