+++ gazon-service/src/main/java/com/gitlab/repository/BankCardRepository.java
@@ -1,7 +1,6 @@
 package com.gitlab.repository;
 
 import com.gitlab.model.BankCard;
-import lombok.NonNull;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -14,12 +13,10 @@ import java.util.List;
 public interface BankCardRepository extends JpaRepository<BankCard, Long> {
 
     @Override
-    @NonNull
     @Query("SELECT u FROM BankCard u order by u.id asc")
     List<BankCard> findAll();
 
     @Override
-    @NonNull
     @Query("SELECT u FROM BankCard u order by u.id asc")
     Page<BankCard> findAll(Pageable pageable);
 }