+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - #task_number - short task description
 - #200 - move search logic to search-service
 - #196 - Add User to Review
+- #173 - Add pagination to ProductSearchController
+- #194 - Add User to Review
 - #136 - Checking functionality ReviewImage Rest
 - #146 - created a log starter
 - #185 - Add Spotbugs
@@ -12,6 +14,8 @@
 - #193 - Checking and add uniqueness_restrictions in the database
 - #198 - Add sorting of payments by ID
 - #175 - Updated security
+- #204 - Add a Central Bank client
+- #183 - Add Security BankCard
 
 ## [1.4.0] - 2.04.2024
 - #167 - Testing the conversion of Product review to productDTO rating